Collapse with specific focus (does not work)
This version now knows if focus is within specific sections of the nav (not all of them). The current biggest issue is that when tabbing while the About section is expanded, it will automatically close the section. There is nothing that states this should happen. You can expand it with enter, and escape to close it. But tab does not currently go onward.